IO Abstraction (Phase 8): Wrap up, clean & finalize
Last of the change to abstract IO and redirect it to FileSystem.

* Split the FileSystem implementation into two - FileSystem & PlatformFileSystem
  FileSystem is pure abstract and has no concept of std::filesystem::path.
  PlatformFileSystem, however, is layer between Surelog & native/OS file system.
* Improve on API documentation
* Enforced naming convention in the API to be explicit
