+++ src/DesignCompile/UhdmChecker.cpp
@@ -39,7 +39,6 @@
 #include <uhdm/uhdm.h>
 #include <uhdm/vpi_visitor.h>
 
-#include <filesystem>
 #include <iomanip>
 #include <sstream>
 #include <stack>
@@ -260,7 +259,7 @@ bool UhdmChecker::reportHtml(PathId uhdmFileId, float overallCoverage) {
       fileSystem->close(report);
       return false;
     }
-    const std::filesystem::path filepath = fileSystem->toPath(fC->getFileId());
+    const std::string filepath(fileSystem->toPath(fC->getFileId()));
     const PathId chkFileId =
         fileSystem->getCheckerHtmlFile(uhdmFileId, fileIndex, symbols);
     const std::string fname(
@@ -286,22 +285,22 @@ bool UhdmChecker::reportHtml(PathId uhdmFileId, float overallCoverage) {
         "#82E0AA; margin:0; min-width: 110px; padding:10; float: left; \">" +
         coverage +
         "</h3> <h3 style=\"margin:0; padding:10; float: left; \"> <a href=" +
-        fname + "> " + filepath.string() + "</a></h3></div>\n";
+        fname + "> " + filepath + "</a></h3></div>\n";
     const std::string fileStatPink =
         "<div style=\"overflow: hidden;\"> <h3 style=\"background-color: "
         "#FFB6C1; margin:0; min-width: 110px; padding:10; float: left; \">" +
         coverage +
         "</h3> <h3 style=\"margin:0; padding:10; float: left; \"> <a href=" +
-        fname + "> " + filepath.string() + "</a></h3></div>\n";
+        fname + "> " + filepath + "</a></h3></div>\n";
     const std::string fileStatRed =
         "<div style=\"overflow: hidden;\"> <h3 style=\"background-color: "
         "#FF0000; margin:0; min-width: 110px; padding:10; float: left; \">" +
         coverage +
         "</h3> <h3 style=\"margin:0; padding:10; float: left; \"> <a href=" +
-        fname + "> " + filepath.string() + "</a></h3></div>\n";
+        fname + "> " + filepath + "</a></h3></div>\n";
     const std::string fileStatWhite =
         "<h3 style=\"margin:0; padding:0 \"> <a href=" + fname + ">" +
-        filepath.string() + "</a> " + coverage + "</h3>\n";
+        filepath + "</a> " + coverage + "</h3>\n";
 
     reportF << "<h3>" << filepath << coverage << "</h3>\n";
     bool uncovered = false;
