+++ src/ErrorReporting/Report.cpp
@@ -142,9 +142,9 @@ std::pair<bool, bool> Report::makeDiffCompUnitReport(CommandLineParser* clp,
 
   const PathId allCompileDirId = fileSystem->getCompileDir(false, st);
   const PathId unitCompileDirId = fileSystem->getCompileDir(true, st);
-  const fs::path allCompileDir = fileSystem->toPath(allCompileDirId);
-  const fs::path unitCompileDir = fileSystem->toPath(unitCompileDirId);
-  const fs::path diffFile = fileSystem->toPath(diffFileId);
+  const fs::path allCompileDir = fileSystem->toPlatformPath(allCompileDirId);
+  const fs::path unitCompileDir = fileSystem->toPlatformPath(unitCompileDirId);
+  const fs::path diffFile = fileSystem->toPlatformPath(diffFileId);
 
   std::string diffCmd = StrCat("diff -r ", unitCompileDir, " ", allCompileDir,
                                " --exclude cache --brief > ", diffFile);
