+++ src/SourceCompile/AntlrParserErrorListener.cpp
@@ -49,8 +49,8 @@ void AntlrParserErrorListener::syntaxError(
     if (!lineText.empty()) {
       lineText.push_back('\n');
       lineText.append(charPositionInLine, ' ');
-      StrAppend(&lineText, "^-- ", fileSystem->toSymbol(m_fileId), ":", line,
-                ":", charPositionInLine, ":");
+      StrAppend(&lineText, "^-- ", fileSystem->toPath(m_fileId), ":", line, ":",
+                charPositionInLine, ":");
     }
   }
   if (m_reportedSyntaxError == false) {
